Keep test Rails app directory clean
All tests are run against a Rails app. The app is generated before each
test, and one would think it gets removed after each test. For unit
tests, this *does* happen before each test, but for acceptance tests, it
does not get removed at all. This commit ensures that this happens, so
that it is possible to run acceptance tests immediately after running
unit tests.