Extract classes for defining models in tests
The main driver behind this commit is to provide a programmatic way to
define models in tests. We already have ways of doing this, of course,
with `define_model` and `define_active_model_class`, but these methods
are very low-level, and in writing tests we have historically made our
own methods inside of test files to define full and complete models. So
we have this common pattern of defining a model with a validation, and
that's repeated across many different files.

What we would like to do, right now, is extract some commonly used
assertions to a shared example group. These assertions need to define
models inside of the tests, but the issue is that sometimes the models
are ActiveRecord models, and sometimes they are ActiveModel models, and
when the shared example group is used within a test file, we need a way
to choose the strategy we'd like to use at runtime. Since the way we
currently define models is via methods, we can't really provide a
strategy very easily. Also, if we need to customize how those models are
defined (say, the attribute needs to be a has-many association instead
of a normal attribute) then the methods only go so far in providing us
that level of customization before things get really complicated.

So, to help us with this, this commit takes the simple pattern of
model-plus-validation previously mentioned and places it in a class,
then splits up that logic. So there are actually four main classes that
we are introducing here:

* CreateModel: for creating a model in an abstract way. You'll couple
  this with one of two lower-level strategies, which may also be used on
  their own:
  * ModelCreationStrategies::ActiveModel
  * ModelCreationStrategies::ActiveRecord
* ActiveRecord::CreateTable: for creating a table that goes along with
  an ActiveRecord model.

Note that `define_model` and `define_active_model_class` have not been
changed so as not to completely break every single test, but they now
simply call out to ModelCreationStrategies::ActiveModel and
ModelCreationStrategies::ActiveRecord internally.